Give the audio format fixture its own module
output_format_str is derived by the server, so the fixture derives it
too instead of sending an empty string it never sends.
